Move rendering to be per-composition-layer
