+++ examples/composition_layer_quad/CMakeLists.txt
@@ -10,10 +10,10 @@ set( SOURCES
   interactions/interaction_teleport.cpp
 )
 
-add_executable( vsgvr_example_skybox ${SOURCES} )
-target_include_directories( vsgvr_example_skybox PRIVATE . )
-target_link_libraries( vsgvr_example_skybox vsg::vsg vsgvr )
+add_executable( vsgvr_example_composition_quad ${SOURCES} )
+target_include_directories( vsgvr_example_composition_quad PRIVATE . )
+target_link_libraries( vsgvr_example_composition_quad vsg::vsg vsgvr )
 
-install(TARGETS vsgvr_example_skybox)
+install(TARGETS vsgvr_example_composition_quad)
 
-set_target_properties( vsgvr_example_skybox PROPERTIES FOLDER "examples" )
+set_target_properties( vsgvr_example_composition_quad PROPERTIES FOLDER "examples" )
