+++ vsgvr/include/vsgvr/app/Viewer.h
@@ -37,7 +37,6 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include <vsgvr/app/CompositionLayers.h>
 
 #include <vsg/app/CommandGraph.h>
-#include <vsg/app/RecordAndSubmitTask.h>
 #include <vsg/app/CompileManager.h>
 #include <vsg/app/UpdateOperations.h>
 #include <vsg/app/RenderGraph.h>
@@ -72,9 +71,9 @@ namespace vsgvr {
             Viewer() = delete;
             ~Viewer();
 
-            // TODO: Update this - Summary level of what to do, based on the event updates.
-            //       Simple things don't need any input from app, so it might just reduce
-            //       to a boolean good/exit status.
+            vsg::ref_ptr<Instance> getInstance() { return _instance; }
+            vsg::ref_ptr<Session> getSession() { return _session; }
+            vsg::ref_ptr<Traits> getTraits() { return _xrTraits; }
 
             /**
              * Whether polling for events succeeded, and what action the application
@@ -123,33 +122,7 @@ namespace vsgvr {
             XrFrameState frameState() const { return _frameState; }
             vsg::ref_ptr<vsg::FrameStamp> frameStamp() const { return _frameStamp; }
 
-            // Manage the work to do each frame using RecordAndSubmitTasks
-            // the vsgvr::Viewer renders to a series of OpenXR composition layers,
-            // each of which contains one or more RecordAndSubmitTasks (for each view/eye)
-            // Composition layers are rendered in order, following the OpenXR specification.
-            
-            // TODO: Some kind of vector of <composition layer definition, list<RecordAndSubmitTasks>
-            // - A base class for CompositionLayer - For any composition layer type, with direct access to struct if needed (to allow extensions)
-            // - Child classes for XrCompositionLayerProjection and XrCompositionLayerQuad
-            // - Optionally child classes for XrCompositionLayerCubeKHR and friends, though as those are extensions would need to be special / marked as such
-            //   - If classes are in vsgvr which require extensions, should be in an extension namespace? or have a common BaseClass::extensionsRequired function?
-            // - Here, a list of structs, pairing a composition layer together with RecordAndSubmitTasks
-            // - In rendering, run each of the tasks in turn to render to the views, and provide those views to OpenXR through the xrEndFrame / composition layers
-            using RecordAndSubmitTasks = std::vector<vsg::ref_ptr<vsg::RecordAndSubmitTask>>;
-            struct CompositionLayerTask
-            {
-              vsg::ref_ptr<vsgvr::CompositionLayer> layer;
-              RecordAndSubmitTasks recordAndSubmitTasks;
-            };
-            std::vector<CompositionLayerTask> compositionRecordAndSubmitTasks;
-
-            // TODO: These methods are required at the moment, and have some small differences from their vsg
-            //       counterparts. Ideally these would not be duplicated, but this will likely require chnanges
-            //       within vsg to correct. In the long run Viewer should only be concerned with the XR parts
-            //       or may even be moved to be a 'Window' class.
-            vsg::CommandGraphs createCommandGraphsForView(vsg::ref_ptr<vsgvr::CompositionLayer> compositionLayer, vsg::ref_ptr<vsg::Node> vsg_scene, std::vector<vsg::ref_ptr<vsg::Camera>>& cameras, bool assignHeadlight = true);
-            void assignRecordAndSubmitTask(vsg::ref_ptr<vsgvr::CompositionLayer> compositionLayer, std::vector<vsg::ref_ptr<vsg::CommandGraph>> in_commandGraphs);
-            void compile(vsg::ref_ptr<vsg::ResourceHints> hints = {});
+            std::vector<vsg::ref_ptr<vsgvr::CompositionLayer>> compositionLayers;
 
             // OpenXR spaces, which will be synced along with actions
             // Typically used for obtaining the position of the headset, via
@@ -162,14 +135,8 @@ namespace vsgvr {
             // One or more may be active at a time, depending on user interaction mode
             std::vector<vsg::ref_ptr<ActionSet>> activeActionSets;
 
-            double nearPlane = 0.05;
-            double farPlane = 100.0;
-
         private:
-            vsg::ref_ptr<vsg::Camera> createCamera(const VkExtent2D& extent);
-
             void shutdownAll();
-            void getViewConfiguration();
 
             void syncSpaceBindings();
             void syncActions();
@@ -182,9 +149,6 @@ namespace vsgvr {
             void createActionSpacesAndAttachActionSets();
             void destroySpaceBindings();
             void destroyActionSpaces();
-
-            void renderCompositionLayerProjection(vsg::ref_ptr<vsgvr::CompositionLayerProjection> layer, RecordAndSubmitTasks& recordAndSubmitTasks);
-            void renderCompositionLayerQuad(vsg::ref_ptr<vsgvr::CompositionLayerQuad> layer, RecordAndSubmitTasks& recordAndSubmitTasks);
 
             vsg::ref_ptr<Instance> _instance;
             vsg::ref_ptr<Traits> _xrTraits;
